@@ -3,7 +3,7 @@ module AcceptsLocationParams
 
   def location_params(root_params)
     tmp_params = root_params.require(:location).permit([:type,{:coordinates => []}])
-    {location: RGeo::GeoJSON.decode(tmp_params).as_text}
+    {location: RGeo::GeoJSON.decode(tmp_params.as_json).as_text}
   end
 
 end
